% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CreateModelList}
\alias{CreateModelList}
\title{Use S3 class as container for model objects (ModelList)}
\usage{
CreateModelList(...)
}
\arguments{
\item{...}{all generated models that you want to add to the class, separated by a comma}
}
\value{
an S3 class ModelList that contains each model input as a parameter
}
\description{
\code{CreateModelList()} takes multiple models as input and creates an S3
ModelList class as a container of models that can be used for predicting and summarizing.
}
